Introduced gulp-based compilation system
