+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
 .DS_Store
 /coverage
+dist
+zones-compiled.js
