+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "ical-expander",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "ICS / iCal / iCalendar parser / expander",
-  "main": "index.js",
+  "main": "dist/ical-expander.js",
   "scripts": {
     "test": "env TZ=Etc/UTC mocha && eslint .",
     "coveralls": "istanbul cover _mocha && istanbul-coveralls"
@@ -30,10 +30,12 @@
     "eslint-plugin-import": "^2.2.0",
     "istanbul": "^0.4.5",
     "istanbul-coveralls": "^1.0.3",
-    "mocha": "^3.2.0"
+    "mocha": "^3.2.0",
+	"gulp": "^3.9.1",
+	"gulp-change": "^1.0.0",
+	"gulp-rename": "^1.2.2"
   },
   "files": [
-    "index.js",
-    "zones-compiled.json"
+    "dist/ical-expander.js"
   ]
 }
